refactor: split chat moderation and scope private room updates
## Summary
- move moderation queue and ban handling into ChatModerationService
- keep ChatService focused on chat posting, reading, and reporting
- route private room events only to joined game-instance sockets while keeping public room events global

## Reasons for Change
- moderation endpoints should not depend on general chat service behavior
- private room edits should stay visible to authorized room members instead of being dropped
