+++ apps/api/src/app/chat/chat.module.ts
@@ -7,10 +7,11 @@ import { UserProfilesModule } from "../user-profiles/user-profiles.module";
 import { AuthModule } from "../../auth/auth.module";
 import { ModerationController } from "./moderation.controller";
 import { ProbableWaffleModule } from "../probable-waffle/probable-waffle.module";
+import { ChatModerationService } from "./chat-moderation.service";
 
 @Module({
   imports: [UserProfilesModule, AuthModule, forwardRef(() => ProbableWaffleModule)],
-  providers: [ChatService, TextSanitizationService, ChatGateway],
+  providers: [ChatService, ChatModerationService, TextSanitizationService, ChatGateway],
   controllers: [ChatController, ModerationController],
   exports: [ChatService]
 })
