+++ apps/api/src/app/chat/chat.service.interface.ts
@@ -1,12 +1,8 @@
 import { type AuthUser } from "@supabase/supabase-js";
 import type {
-  BanUserDto,
   ChatMessage,
   GetMessagesResponseDto,
-  ModerationQueueDto,
-  ModerationSummaryDto,
-  ReportChatMessageDto,
-  UpdateChatReportStatusDto
+  ReportChatMessageDto
 } from "@fuzzy-waddle/api-interfaces";
 
 export interface IChatService {
@@ -18,9 +14,4 @@ export interface IChatService {
     user: AuthUser
   ): Promise<GetMessagesResponseDto>;
   reportMessage(messageId: number, user: AuthUser, report: ReportChatMessageDto): Promise<void>;
-  getModerationSummary(user: AuthUser): Promise<ModerationSummaryDto>;
-  getModerationReports(user: AuthUser): Promise<ModerationQueueDto>;
-  updateReportStatus(reportId: number, user: AuthUser, body: UpdateChatReportStatusDto): Promise<void>;
-  banUser(userId: string, user: AuthUser, body: BanUserDto): Promise<void>;
-  unbanUser(userId: string, user: AuthUser): Promise<void>;
 }
