+++ apps/api/src/app/chat/moderation.controller.ts
@@ -3,22 +3,22 @@ import type { AuthUser } from "@supabase/supabase-js";
 import type { ModerationQueueDto, ModerationSummaryDto } from "@fuzzy-waddle/api-interfaces";
 import { CurrentUser } from "../../auth/current-user";
 import { ModeratorAccessGuard } from "../../auth/guards/moderator-access.guard";
-import { ChatService } from "./chat.service";
+import { ChatModerationService } from "./chat-moderation.service";
 import { BanUserBodyDto, UpdateChatReportStatusBodyDto } from "./report-chat-message.dto";
 
 @Controller("moderation")
 @UseGuards(ModeratorAccessGuard)
 export class ModerationController {
-  constructor(private readonly chatService: ChatService) {}
+  constructor(private readonly chatModerationService: ChatModerationService) {}
 
   @Get("summary")
   async getModerationSummary(@CurrentUser() user: AuthUser): Promise<ModerationSummaryDto> {
-    return this.chatService.getModerationSummary(user);
+    return this.chatModerationService.getModerationSummary(user);
   }
 
   @Get("reports")
   async getModerationReports(@CurrentUser() user: AuthUser): Promise<ModerationQueueDto> {
-    return this.chatService.getModerationReports(user);
+    return this.chatModerationService.getModerationReports(user);
   }
 
   @Post("reports/:reportId/status")
@@ -27,7 +27,7 @@ export class ModerationController {
     @Param("reportId", ParseIntPipe) reportId: number,
     @Body() body: UpdateChatReportStatusBodyDto
   ): Promise<void> {
-    return this.chatService.updateReportStatus(reportId, user, body);
+    return this.chatModerationService.updateReportStatus(reportId, user, body);
   }
 
   @Post("users/:userId/ban")
@@ -36,11 +36,11 @@ export class ModerationController {
     @Param("userId") userId: string,
     @Body() body: BanUserBodyDto
   ): Promise<void> {
-    return this.chatService.banUser(userId, user, body);
+    return this.chatModerationService.banUser(userId, user, body);
   }
 
   @Post("users/:userId/unban")
   async unbanUser(@CurrentUser() user: AuthUser, @Param("userId") userId: string): Promise<void> {
-    return this.chatService.unbanUser(userId, user);
+    return this.chatModerationService.unbanUser(userId, user);
   }
 }
