+++ apps/api/src/app/probable-waffle/game-room/room.gateway.ts
@@ -1,5 +1,9 @@
 import { OnGatewayConnection, WebSocketGateway, WebSocketServer } from "@nestjs/websockets";
-import { ProbableWaffleGatewayEvent, type ProbableWaffleRoomEvent } from "@fuzzy-waddle/api-interfaces";
+import {
+  ProbableWaffleGatewayEvent,
+  ProbableWaffleGatewayRoomTypes,
+  type ProbableWaffleRoomEvent
+} from "@fuzzy-waddle/api-interfaces";
 import { Server, Socket } from "socket.io";
 import { SocketConnectionAuthService } from "../../../auth/socket-connection-auth.service";
 
@@ -19,5 +23,11 @@ export class RoomGateway implements OnGatewayConnection {
 
   emitRoom(roomEvent: ProbableWaffleRoomEvent) {
     this.server.emit(ProbableWaffleGatewayEvent.ProbableWaffleRoom, roomEvent);
+  }
+
+  emitRoomToGameInstance(gameInstanceId: string, roomEvent: ProbableWaffleRoomEvent) {
+    this.server
+      .to(`${ProbableWaffleGatewayRoomTypes.ProbableWaffleGameInstance}${gameInstanceId}`)
+      .emit(ProbableWaffleGatewayEvent.ProbableWaffleRoom, roomEvent);
   }
 }
